backend/management_ms/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
mysql-connector-python
requests